Add --enable-valgrind flag to configure script.
If you configure with this script, valgrind is run on every test.
This will make the tests take 10-20x more time, and it will make every
test that has memory-leaks fail. You might not want them to fail, and if
that is the case then edit the valgrind-wrapper script to return 0 even
if there are leaks in the tests. This way the tests are allowed to pass
even if they contain leaks :)